Using type_traits instead of point_traits
