+++ common/CMakeLists.txt
@@ -60,6 +60,7 @@ set(incs
   include/pcl/types.h
   include/pcl/point_cloud.h
   include/pcl/point_traits.h
+  include/pcl/type_traits.h
   include/pcl/point_types_conversion.h
   include/pcl/point_representation.h
   include/pcl/point_types.h
