+++ common/include/pcl/point_cloud.h
@@ -48,7 +48,7 @@
 #include <pcl/exceptions.h>
 #include <pcl/memory.h>
 #include <pcl/pcl_macros.h>
-#include <pcl/point_traits.h>
+#include <pcl/type_traits.h>
 
 #include <algorithm>
 #include <utility>
