+++ surface/include/pcl/surface/impl/mls.hpp
@@ -40,7 +40,7 @@
 #ifndef PCL_SURFACE_IMPL_MLS_H_
 #define PCL_SURFACE_IMPL_MLS_H_
 
-#include <pcl/point_traits.h>
+#include <pcl/type_traits.h>
 #include <pcl/surface/mls.h>
 #include <pcl/common/io.h>
 #include <pcl/common/copy_point.h>
