+++ test/common/test_type_traits.cpp
@@ -37,7 +37,7 @@
 
 
 #include <pcl/memory.h>  // for pcl::has_custom_allocator, PCL_MAKE_ALIGNED_OPERATOR_NEW
-#include <pcl/point_traits.h>
+#include <pcl/type_traits.h>
 #include <pcl/point_types.h>
 
 #include <pcl/test/gtest.h>
